pyproject.toml
[tool.poetry]
name = "image-transform"
version = "0.1.0"
description = ""
authors = ["Taras Nytiaha <taras.nytiaha@gmail.com>"]
readme = "README.md"
packages = [{include = "image_transform"}]

[tool.poetry.dependencies]
python = "3.11.0"
Django = "^5.1.1"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
